re-ID line: WHICH-under-reacquisition fix (crop-DINOv2 + K2 gallery + confidence gate)
Closed-loop re-acquisition研究线: reframe SR wall from WHERE to WHICH, then a
training-free / resolution-free fix that lifts deployable SR from 0 to 0.105
(≈ expert-control ceiling), significant mis_follow 0.70→0.40 & lock_wrong 0.74→0.49.

Harness (carla_uav_tracking/rollout/): road_predictor + ex_source road/track,
tid_head=reid (crop-DINOv2 feature + K-view diversity gallery + top-m match),
confidence-gated control (conf_tau); reid_resolution_probe (feature/max_pixels/bank sweep);
baseline_deadreckon road branch. Result jsons under rollout/eval_out.

Docs (acot_note/): reid-line-summary (paper spine), reid-qa-faq (2026-09-03),
SVG pipeline + final-architecture diagrams, wm-p2a-plan (WM closed: WHERE not the
bottleneck), reid-highres-render-spec, findings-synthesis, research-bottlenecks.

+++ train/backbone_kv.py
@@ -130,7 +130,7 @@ def precompute_multilayer(cfg, layers, shard=0, num_shards=1, device=None):
                     per_frame.append(np.stack([_pool_grid(feats[L][0], g) for L in used]))
             if not per_frame or used is None:
                 continue
-            ctx_arr = np.stack(per_frame).astype(np.float32)
+            ctx_arr = np.stack(per_frame).astype(np.float16)         # fp16: half disk, VLM feats tolerate it
             np.save(out.with_name(out.stem + ".ctx.npy"), ctx_arr)   # uncompressed -> mmap-ready
             np.savez(out, frames=np.array(frames, np.int64),         # slim meta (no ctx)
                      grid=np.array([g, g], np.int64),
